+++ src/Spendings.java
@@ -1,3 +1,4 @@
+import java.math.BigDecimal;
 import java.util.GregorianCalendar;
 
 /**
@@ -8,13 +9,13 @@ import java.util.GregorianCalendar;
  */
 public class Spendings extends Finances {
 	
-	public Spendings(double expenditure, GregorianCalendar date) {
+	public Spendings(BigDecimal expenditure, GregorianCalendar date) {
 		// income should be positive, valid date
 		super(expenditure, date);
 		// returns Spendings instance
 	}
 	
-	public double getExpenditure() {
+	public BigDecimal getExpenditure() {
 		// amount has to be initialized
 		return this.amount;
 		// returns stored amount
